[feature ds-better-adapter-populated-record-array-error-messages]
The result returned from `store.query` is an
AdapterPopulatedRecordArray. This adds better error messages when you
try to mutate the array, as it is considered immutable and the mutation
methods should not be used.

Currently, there are some issues like #3886 where this does not fail in
a helpful way. This feature improves the developer ergonomics to prevent
the mutation from happening and explaining why.